requirement.txt
fastapi
uvicorn
jinja2
pillow